+++ eslint.config.js
@@ -7,7 +7,7 @@ import prettier from "eslint-plugin-prettier";
 import tseslint from "typescript-eslint";
 
 export default tseslint.config(
-  { ignores: ["dist", "lighthouse-with-auth.js"] },
+  { ignores: ["dist", "scripts/lighthouse-with-auth.cjs"] },
   {
     files: ["eslint.config.js", "*.js"],
     ...js.configs.recommended,
